fix json serialization
