+++ packages/models-library/src/models_library/utils/_original_fastapi_encoders.py
@@ -3,92 +3,21 @@
 # wget https://raw.githubusercontent.com/tiangolo/fastapi/master/fastapi/encoders.py --output-document=_original_fastapi_encoders
 #
 import dataclasses
-import datetime
 from collections import defaultdict, deque
-from decimal import Decimal
 from enum import Enum
-from ipaddress import (
-    IPv4Address,
-    IPv4Interface,
-    IPv4Network,
-    IPv6Address,
-    IPv6Interface,
-    IPv6Network,
-)
-from pathlib import Path, PurePath
-from re import Pattern
+from pathlib import PurePath
 from types import GeneratorType
 from typing import Any, Callable, Union
-from uuid import UUID
 
+from models_library.utils.json_serialization import ENCODERS_BY_TYPE
 from pydantic import BaseModel
-from pydantic.color import Color
-from pydantic.networks import AnyUrl, NameEmail
-from pydantic.types import SecretBytes, SecretStr
-from pydantic_core import PydanticUndefined, PydanticUndefinedType, Url
+from pydantic_core import PydanticUndefined, PydanticUndefinedType
 from typing_extensions import Annotated, Doc
 
 Undefined = PydanticUndefined
 UndefinedType = PydanticUndefinedType
 
 IncEx = Union[set[int], set[str], dict[int, Any], dict[str, Any]]
-
-# Taken from Pydantic v1 as is
-def isoformat(o: datetime.date | datetime.time) -> str:
-    return o.isoformat()
-
-
-# Taken from Pydantic v1 as is
-# TODO: pv2 should this return strings instead?
-def decimal_encoder(dec_value: Decimal) -> int | float:
-    """
-    Encodes a Decimal as int of there's no exponent, otherwise float
-
-    This is useful when we use ConstrainedDecimal to represent Numeric(x,0)
-    where a integer (but not int typed) is used. Encoding this as a float
-    results in failed round-tripping between encode and parse.
-    Our Id type is a prime example of this.
-
-    >>> decimal_encoder(Decimal("1.0"))
-    1.0
-
-    >>> decimal_encoder(Decimal("1"))
-    1
-    """
-    if dec_value.as_tuple().exponent >= 0:  # type: ignore[operator]
-        return int(dec_value)
-    else:
-        return float(dec_value)
-
-
-ENCODERS_BY_TYPE: dict[type[Any], Callable[[Any], Any]] = {
-    bytes: lambda o: o.decode(),
-    Color: str,
-    datetime.date: isoformat,
-    datetime.datetime: isoformat,
-    datetime.time: isoformat,
-    datetime.timedelta: lambda td: td.total_seconds(),
-    Decimal: decimal_encoder,
-    Enum: lambda o: o.value,
-    frozenset: list,
-    deque: list,
-    GeneratorType: list,
-    IPv4Address: str,
-    IPv4Interface: str,
-    IPv4Network: str,
-    IPv6Address: str,
-    IPv6Interface: str,
-    IPv6Network: str,
-    NameEmail: str,
-    Path: str,
-    Pattern: lambda o: o.pattern,
-    SecretBytes: str,
-    SecretStr: str,
-    set: list,
-    UUID: str,
-    Url: str,
-    AnyUrl: str,
-}
 
 
 def generate_encoders_by_class_tuples(
